requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
sqlmodel==0.0.21
easyocr==1.7.2
pillow==10.4.0
python-multipart==0.0.9
python-dotenv==1.0.1
